CORE-20203: Move some common code into external event responses
