+++ components/ledger/ledger-verification/src/main/kotlin/net/corda/ledger/verification/processor/impl/VerificationExceptionCategorizerImpl.kt
@@ -3,6 +3,7 @@ package net.corda.ledger.verification.processor.impl
 import net.corda.data.flow.event.external.ExternalEventResponseErrorType
 import net.corda.flow.external.events.responses.exceptions.CpkNotAvailableException
 import net.corda.flow.external.events.responses.exceptions.NotAllowedCpkException
+import net.corda.flow.external.events.responses.exceptions.criteria
 import net.corda.ledger.verification.processor.VerificationExceptionCategorizer
 import java.io.NotSerializableException
 
@@ -16,24 +17,6 @@ class VerificationExceptionCategorizerImpl : VerificationExceptionCategorizer {
             else -> ExternalEventResponseErrorType.PLATFORM
         }
     }
-
-    private data class ExceptionCriteria<T : Throwable>(val type: Class<T>, val check: (T) -> Boolean = { _ -> true }) {
-        fun meetsCriteria(exception: Throwable?): Boolean {
-            if (exception == null) {
-                return false
-            }
-            val meetsCriteria = if (type.isAssignableFrom(exception::class.java)) {
-                check(type.cast(exception))
-            } else {
-                false
-            }
-            return (meetsCriteria || meetsCriteria(exception.cause))
-        }
-    }
-
-    private inline fun <reified T : Throwable> criteria(
-        noinline check: (T) -> Boolean = { _ -> true }
-    ): ExceptionCriteria<T> = ExceptionCriteria(T::class.java, check)
 
     private fun isPossiblyFatal(exception: Throwable): Boolean {
         val checks = listOf(
